+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/token/domain/TokenType.java
@@ -1,4 +1,4 @@
-package com.playkuround.playkuroundserver.domain.token.domain;
+package com.playkuround.playkuroundserver.domain.auth.token.domain;
 
 public enum TokenType {
