+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/token/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.playkuround.playkuroundserver.domain.token.exception;
+package com.playkuround.playkuroundserver.domain.auth.token.exception;
 
 import com.playkuround.playkuroundserver.global.error.exception.InvalidValueException;
